refactor: Move shared evaluator helper functions into evaluator_common
Both evaluators duplicated ~150 lines of pure, stateless helper functions and constants verbatim; move them (with EvalResult/EvaluationException) to impl/evaluator_common.py imported by both. Drops now-dead imports (hashlib, operators, Any, Set). Touches released sync evaluator.py — behavior identical.
